♻️ Refactor:  채팅 ui 코드 스플리팅 - #QTVY-107
