+++ src/app/(default-layout)/chat/page.tsx
@@ -1,4 +1,4 @@
-import ChatRoomList from '@/features/chat/ui/ChatRoomList';
+import ChatRoomList from '@/features/chat/ui/ChatRoomList/ChatRoomList';
 
 export default function ChatPage() {
   return (
