+++ src/features/chat/types.ts
@@ -1,0 +1,14 @@
+export interface ChatRoom {
+  id: string;
+  name: string;
+  latestMessage: string;
+  latestTimestamp: string;
+  type: 'group' | 'personal';
+  sports: string;
+  status: 'active' | 'completed';
+}
+
+export interface SideModalProps {
+  roomInfo: ChatRoom;
+  onClose: () => void;
+}
