+++ src/features/chat/ui/ChatRoomList/ChatRoomList.tsx
@@ -3,58 +3,13 @@ import { useState } from 'react';
 import Link from 'next/link';
 import { ChatRoomCard } from '@/components/common/Cards/ChatRoomCard';
 import { Toggle } from '@/components/common/Toggle';
-
-export interface ChatRoom {
-  id: string;
-  name: string;
-  latestMessage: string;
-  latestTimestamp: string;
-  type: 'group' | 'personal';
-  sports: string;
-  status: 'active' | 'completed';
-}
+import { ChatData } from './constants';
+import { ChatRoom } from '../../types';
 
 export default function ChatRoomList() {
   const [activeTab, setActiveTab] = useState<'group' | 'personal'>('group');
   const [isActiveChat, setIsActiveChat] = useState<boolean>(true);
-  const [chatRooms] = useState<ChatRoom[]>([
-    {
-      id: '1',
-      name: '농구에 미친 사럼들',
-      latestMessage: '종합운동장에서 하는 거 맞나요??',
-      latestTimestamp: '2024.11.28',
-      type: 'group',
-      sports: '농구',
-      status: 'active'
-    },
-    {
-      id: '2',
-      name: '김철수',
-      latestMessage: '오늘 저녁 러닝 ㄱ?',
-      latestTimestamp: '2024.11.28',
-      type: 'personal',
-      sports: '',
-      status: 'completed'
-    },
-    {
-      id: '3',
-      name: '축구 동호회',
-      latestMessage: '다음 경기 일정 잡자',
-      latestTimestamp: '2024.11.27',
-      type: 'group',
-      sports: '축구',
-      status: 'active'
-    },
-    {
-      id: '4',
-      name: '이영희',
-      latestMessage: '테니스 레슨 끝났어',
-      latestTimestamp: '2024.11.26',
-      type: 'personal',
-      sports: '',
-      status: 'completed'
-    }
-  ]);
+  const [chatRooms] = useState<ChatRoom[]>(ChatData);
 
   const filteredChatRooms = chatRooms.filter(
     room =>
